add new conftest file in z_cluster
